Account for deleting blobs when counting space usage
